refactor: adapted directory structur to new pattern
