+++ internal/tests/R00/ex00.go
@@ -9,10 +9,10 @@ import (
 	"github.com/42-Short/shortinette/internal/errors"
 	"github.com/42-Short/shortinette/internal/functioncheck"
 	"github.com/42-Short/shortinette/internal/logger"
-	"github.com/42-Short/shortinette/internal/testbuilder"
+	"github.com/42-Short/shortinette/internal/interfaces/exercise"
 )
 
-func ex00Compile(test *testbuilder.Test) error {
+func ex00Compile(test *exercisebuilder.Test) error {
 	cmd := exec.Command("rustc", test.TurnInFile)
 	cmd.Dir = fmt.Sprintf("studentcode/%s/", test.TurnInDirectory)
 
@@ -45,7 +45,7 @@ func assertionErrorString(testName string, expected string, got string) string {
 	return fmt.Sprintf("[%s KO]: %v", testName, outputComparison)
 }
 
-func ex00Test(test *testbuilder.Test) bool {
+func ex00Test(test *exercisebuilder.Test) bool {
 	if err := functioncheck.Execute(*test, "shortinette-test-R00"); err != nil {
 		logger.File.Printf("[%s KO]: %v", test.Name, err)
 		return false
@@ -69,8 +69,8 @@ func ex00Test(test *testbuilder.Test) bool {
 	return true
 }
 
-func ex00() testbuilder.TestBuilder {
-	return testbuilder.NewTestBuilder().
+func ex00() exercisebuilder.ExerciseBuilder {
+	return exercisebuilder.NewExerciseBuilder().
 		SetName("EX00").
 		SetTurnInDirectory("ex00").
 		SetTurnInFile("hello.rs").
